Vision/requirements.txt
# langchain_groq
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

langchain
langchain-groq
langchain-core
langsmith